Initial pass, some playtesting. Implemented the UI, moved the content into it's own file, added a new section title with uid and area name.
